v2.9c
morreu mais passa bem - revivi
