+++ pages/perfil-paciente.html
@@ -15,7 +15,7 @@
 
     <script src="/scripts/scroll.js"></script>
     <script src="/scripts/utilites.js" defer></script>
-    <script type="text/javascript" src="/scripts/login_app.js" defer></script>
+    <script type="module" src="/scripts/login_app.js" defer></script>
     <script type="text/javascript" src="/scripts/nav.js" defer></script>
