+++ pages/stream-terapia.html
@@ -10,7 +10,7 @@
     <link rel="stylesheet" href="../bootstrap-4.1.3-dist/css/bootstrap.min.css">
     <script type="text/javascript" src="../bootstrap-4.1.3-dist/js/bootstrap.min.js"></script>
     
-    <script type="text/javascript" src="/scripts/login_app.js" defer></script>
+    <script type="module" src="/scripts/login_app.js" defer></script>
 
     <script>
         usuario = JSON.parse(sessionStorage.getItem('user'))
