+++ scripts/login_app.js
@@ -1,6 +1,17 @@
 const PORT = 8080; // port principal
 
-import { getPacientes } from "../server/db/models/defaultUsers";
+// os pacientes
+export function getPacientes(){
+    const ds = [ 
+        { id: 1, login: "gabriel", password: "1234", email: "gabriel@gmail.com", img_perfil:''},//[0]
+        { id: 2, login: "amanda", password: "12345@", email: "amanda@gmail.com", img_perfil:'amanda.png'},//[1]
+        { id: 3, login: "ladygaga", password: "123456@", email: "ladygaga@gmail.com", img_perfil:''},//[2]
+        { id: 4, login: "snoopy", password: "1950", email: "snoopy@gmail.com", img_perfil:'snoopy.png'},
+        { id: 5, login: "scooby", password: "1950", email: "scooby@gmail.com", img_perfil: 'scooby.png' },
+    ]
+    return ds
+}
+
 function local(){
     if (window.location.hostname.includes("vercel")){
     // quando não tiver o vetor usuario e o site estiver offline
@@ -13,6 +24,7 @@ function local(){
         }
     }
 }
+
 
 function doLogin(event){
     event.preventDefault();
@@ -170,3 +182,13 @@ function getUserTypeColor(){
         loginBody.style.boxShadow  = `inset 0px 14px 0px ${tableColors[userType]}`
     }
 }
+
+window.local = local; 
+window.doLogin = doLogin; 
+window.cadastrar = cadastrar; 
+window.changePerfilData = changePerfilData;
+window.logout = logout;
+window.setTerapia = setTerapia;
+window.getTerapia = getTerapia;
+window.setUserType = setUserType;
+window.getUserTypeColor = getUserTypeColor;
